svn r1114: Converted functions in util to have osd_
i.e. "error()" => "osd_error()". This was because error()
conflicts with libc's error() function.

Note that I realize this change might not be agreed on by all,
so if you don't like it feel free to roll it back (we could also
use #define trickery but that seems like a mess)

SVN Date: 2007-02-21T22:37:13.287101Z
Auto-commited-by: ./svn_fetch3.sh
Squashed-with-old-git: b4294ae
Log-edited-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>