osd-target: extension: Let in a system_ID set on LUN format
system_ID is a, factory set, unique identifier of an osd-lun.
Since osd-target is a software solution we should enable
a system administrator setting of the system_id.

The solution proposed here is to let a system_id attribute-
set while an OSD_FORMAT is preformed on the osd-lun. Rational
is: since we are initializing a new backing storage to host
the osd-lun we can at this time set it's name, and by that
lock it permanently until a new format initializes a new osd.

Thous, regular osd utilities can be used to set the system_id
and no special command line or config files magic is needed.

The system_id set can be up to RIAP_OSD_SYSTEM_ID_LEN bytes in
size, and if smaller is zero left-padded.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>