{TODO} osd-target: up: Launch a 2nd LUN
For multy-device mirror configuration and all future multi device configurations
we should be able to export multiple targets/luns from the same server. Each with
it's own backing store.

TODO: Just us we establish the second lun in this patch, we should remove the automatic-
first-lun patch on /tmp/$user/... and have backing-device path here inside the script.
(And not hard coded in source-code.)

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>