open-osd: Remove of OSC's osd-initiator
The osd-initiator is no longer supported since the git tag
"CDB_VER_OSD2r01". After that we changed target's wire-format
to OSD2r05.
One needs to checkout:
	[]$ git checkout CDB_VER_OSD2r01
for the last version that supported both initiator and target
from the OSC project.

For a user-mode as well as Kernel mode osd initiator that uses
the same exact API checkout:
[]$ git-clone git://git.open-osd.org/open-osd.git

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>