osd-target: Support osd2r04 implementation of scatter/gather lists.
The original OSC implementation used "DDT" bits in the CDB to
enable scatter/gather.  With the introduction of continuations in
osd2r04, scatter/gather was changed to use continuations instead of
the "DDT" bits.

The code in osd.c still supports the OSC vector/strided mode.
However, cdb.c removes the vector/strided support to be compliant with
osd2r04.

In addition, the original OSC code had added scatter/gather support to
the APPEND command.  This has been removed to be compliant with
osd2r04.

Signed-off-by: John A. Chandy <john.chandy@uconn.edu>